Version 1 of the project only the volume of the songs is being iterated
In these files, the project is able to select the mode of the program and then choose the songs accordingly. After the song is chosen, the calibration is done and then the audio is played and then only the volume of the song can be controlled because the audio is not yet preprocessed to control the speed.